+++ CHANGELOG.md
@@ -12,6 +12,7 @@ use patch releases for compatibility fixes instead.
 ### Added
 
 - Added missing `user_restrictions` field to `Cluster` data model.
+- Added `Beaker.job.preempt()` method.
 - Added `Job.was_preempted` property.
 - Added `job` attribute to `JobFailedError` and `task` attribute to `TaskStoppedError`.
 - Added DEBUG logging statements for every request and response to/from the Beaker server. To see these, just enable logging at the DEBUG level (though you may want to disable DEBUG logging from the "urllib3" logger, as that will create a lot of noise). For example: